Refactor get keys related code
